// lireddit-web/src/enums/navigationEnum.ts
export enum FeedsEnum {
  ALL_POSTS = "All",
  POPULAR = "Popular",
  HAPPENING_NOW = "Happening Now",
  HOME = "Home",
  MY_POSTS = "My Posts",
}

export enum OthersEnum {
  USER_SETTINGS = "User Settings",
  MESSAGES = "Messages",
  NOTIFICATIONS = "Notifications",
  CREATE_POST = "Create Post",
  MOST_LIKED = "Most Liked",
  MOST_DISLIKED = "Most Disliked",
  COINS = "Coins",
}